+++ xinu-fall2018/system/myprogA.c
@@ -36,7 +36,7 @@ process myprogA(void)
     asm volatile ("movl %%esp, %0\n\t"
                 : "=r" (esp));
 
-    kprintf("Process Name: %s\n", (uint32)proctab[pid].prname);
+    kprintf("Process Name: %s\n", proctab[pid].prname);
     kprintf("Stack Base: [0x%08X]\n", (uint32)proctab[pid].prstkbase);
     kprintf("Stack Size: %d bytes\n", (uint32)(proctab[pid].prstkbase) - (uint32)(esp) + 4);
     kprintf("Stack Limit: %d bytes\n", proctab[pid].prstklen);
@@ -59,7 +59,7 @@ process myprogA(void)
 
     asm volatile ("movl %%esp, %0\n\t"
                 : "=r" (esp));
-    kprintf("Process Name: %s\n", (uint32)proctab[pid].prname);
+    kprintf("Process Name: %s\n", proctab[pid].prname);
     kprintf("Stack Base: [0x%08X]\n", (uint32)proctab[pid].prstkbase);
     kprintf("Stack Size: %d bytes\n", (uint32)(proctab[pid].prstkbase) - (uint32)(esp) + 4);
     kprintf("Stack Limit: %d bytes\n", proctab[pid].prstklen);
